Move SOCAT-enhanced report generation to programs package
since SOCAT-enhanced reports are not part of the dashboard operation.
(Bundling and sending of original data files still part of the dashboard.)
Allow null to be passed for SST and Salinity in programs.DataPoint.
